apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy[asyncio]
aiosqlite
asyncpg
alembic
redis[hiredis]
celery
python-jose[cryptography]
passlib
bcrypt==3.2.2
pyotp
geoip2
httpx
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp-proto-grpc
# ── Test dependencies ──
pytest
pytest-asyncio

